Register before_delete event for liberty security group deleting
In Liberty, with its AFTER_DELETE hook, there's no way to get the
security group ID to refresh because the information has been lost from
the DB and is not otherwise available.  Add a memory, populated in
BEFORE_DELETE, that will provide the secgroup rule to AFTER_DELETE.

Additionally, clean up the deleted-rules ignorer to clean up rules just
once.

Fixes: LUNAR-1256
Change-Id: I6bdb4037d1f9d022e030f542a6d48d36ab594c72