Remove l2 tag rewrite for vhost and tap interface
VPP 17.01 as of now has a bug: the last L2 tag rewrite status is
remembered for an iface_idx when the idx is re-used.  There's a
chance that a new created vhost/tap interface is on a index that was
a vlan subinterface (or it could have been an interface someone else
created). Here, we forcibly disable l2 tag rewrite for interfaces we
create when there's no other rewrite rule being added to them, in order
to ensure all vhost/tap interface are not rewriting l2 tag.

Change-Id: I6dbea77f6c2057deeaf688b0a9cffed1b0a2b313
Closes-Bug: 1667849