Unbind port properly in update_port_precommit
There are several problems with the update_port_precommit code that
mean that it doesn't properly unbind a port either in the case that
the port changes host or the port is being deleted.

This fixes a range of those cases - in the updating host case, where
"current_bind" will be not None, and in the unbinding host case,
where "binding_levels" will be None. Without this patch we end up with
orphan ports, where etcd has the entry but the Neutron database entry
has been deleted.

Change-Id: Ic97e39f0d9ebac8c8bf546dabb6f8ba579d767e6
Fixes: LUNAR-1255